svg/struct/scripted/svg-getIntersectionList-005.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml"><title>getIntersectionList() on element with intersecting bounding box</title><h:link href="https://svgwg.org/svg2-draft/struct.html#__svg__SVGSVGElement__getIntersectionList" rel="help"/><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><rect id="e0" width="100" height="100" x="110" y="100" fill="red"/><rect id="e1" width="40" height="40" x="160" y="130" fill="green"/><path id="e2" fill="none" stroke="#00f" stroke-width="2" d="M100 200 120 110 220 100"/><script>test(t =&gt; { const svg = document.documentElement; const queryRect = document.getElementById(&apos;e1&apos;).getBBox(); const list = svg.getIntersectionList(queryRect, null); const result = Array.from(list).map(e =&gt; e.id).sort(); assert_array_equals(result, [&apos;e0&apos;, &apos;e1&apos;]); });</script></svg>
